rustc_pattern_analysis::rustc

Function analyze_match

pub fn analyze_match<'p, 'tcx>(
    tycx: &RustcPatCtxt<'p, 'tcx>,
    arms: &[MatchArm<'p, 'tcx>],
    scrut_ty: Ty<'tcx>,
) -> Result<UsefulnessReport<'p, 'tcx>, ErrorGuaranteed>

The entrypoint for this crate. Computes whether a match is exhaustive and which of its arms are useful, and runs some lints.