Struct rustc_pattern_analysis::rustc::RustcPatCtxt

pub struct RustcPatCtxt<'p, 'tcx: 'p> {
    pub tcx: TyCtxt<'tcx>,
    pub typeck_results: &'tcx TypeckResults<'tcx>,
    pub module: DefId,
    pub param_env: ParamEnv<'tcx>,
    pub dropless_arena: &'p DroplessArena,
    pub match_lint_level: HirId,
    pub whole_match_span: Option<Span>,
    pub scrut_span: Span,
    pub refutable: bool,
    pub known_valid_scrutinee: bool,
}

Fields
tcx: TyCtxt<'tcx>typeck_results: &'tcx TypeckResults<'tcx>module: DefId

The module in which the match occurs. This is necessary for checking inhabited-ness of types because whether a type is (visibly) inhabited can depend on whether it was defined in the current module or not. E.g., struct Foo { _private: ! } cannot be seen to be empty outside its module and should not be matchable with an empty match statement.

param_env: ParamEnv<'tcx>dropless_arena: &'p DroplessArena

To allocate the result of self.ctor_sub_tys()

match_lint_level: HirId

Lint level at the match.

whole_match_span: Option<Span>

The span of the whole match, if applicable.

scrut_span: Span

Span of the scrutinee.

refutable: bool

Only produce NON_EXHAUSTIVE_OMITTED_PATTERNS lint on refutable patterns.

known_valid_scrutinee: bool

Whether the data at the scrutinee is known to be valid. This is false if the scrutinee comes from a union field, a pointer deref, or a reference deref (pending opsem decisions).

Implementations

impl<'p, 'tcx: 'p> RustcPatCtxt<'p, 'tcx>

pub fn reveal_opaque_ty(&self, ty: Ty<'tcx>) -> RevealedTy<'tcx>

Type inference occasionally gives us opaque types in places where corresponding patterns have more specific types. To avoid inconsistencies as well as detect opaque uninhabited types, we use the corresponding concrete type if possible.

fn reveal_opaque_key(&self, key: OpaqueTypeKey<'tcx>) -> Option<Ty<'tcx>>

Returns the hidden type corresponding to this key if the body under analysis is allowed to know it.

pub fn is_uninhabited(&self, ty: Ty<'tcx>) -> bool

pub fn is_foreign_non_exhaustive_enum(&self, ty: RevealedTy<'tcx>) -> bool

Returns whether the given type is an enum from another crate declared #[non_exhaustive].

pub fn is_range_beyond_boundaries( &self, range: &IntRange, ty: RevealedTy<'tcx>, ) -> bool

Whether the range denotes the fictitious values before isize::MIN or after usize::MAX/isize::MAX (see doc of IntRange::split for why these exist).

pub(crate) fn variant_sub_tys( &self, ty: RevealedTy<'tcx>, variant: &'tcx VariantDef, ) -> impl Iterator<Item = (&'tcx FieldDef, RevealedTy<'tcx>)> + Captures<'p> + Captures<'_>

pub(crate) fn variant_index_for_adt( ctor: &Constructor<'p, 'tcx>, adt: AdtDef<'tcx>, ) -> VariantIdx

pub(crate) fn ctor_sub_tys<'a>( &'a self, ctor: &'a Constructor<'p, 'tcx>, ty: RevealedTy<'tcx>, ) -> impl Iterator<Item = (RevealedTy<'tcx>, PrivateUninhabitedField)> + ExactSizeIterator + Captures<'a>

Returns the types of the fields for a given constructor. The result must have a length of ctor.arity().

pub(crate) fn ctor_arity( &self, ctor: &Constructor<'p, 'tcx>, ty: RevealedTy<'tcx>, ) -> usize

The number of fields for this constructor.

pub fn ctors_for_ty( &self, ty: RevealedTy<'tcx>, ) -> Result<ConstructorSet<'p, 'tcx>, ErrorGuaranteed>

Creates a set that represents all the constructors of ty.

See crate::constructor for considerations of emptiness.

pub(crate) fn lower_pat_range_bdy( &self, bdy: PatRangeBoundary<'tcx>, ty: RevealedTy<'tcx>, ) -> MaybeInfiniteInt

pub fn lower_pat(&self, pat: &'p Pat<'tcx>) -> DeconstructedPat<'p, 'tcx>

Note: the input patterns must have been lowered through rustc_mir_build::thir::pattern::check_match::MatchVisitor::lower_pattern.

pub(crate) fn hoist_pat_range_bdy( &self, miint: MaybeInfiniteInt, ty: RevealedTy<'tcx>, ) -> PatRangeBoundary<'tcx>

Convert back to a thir::PatRangeBoundary for diagnostic purposes. Note: it is possible to get isize/usize::MAX+1 here, as explained in the doc for IntRange::split. This cannot be represented as a Const, so we represent it with PosInfinity.

pub(crate) fn hoist_pat_range( &self, range: &IntRange, ty: RevealedTy<'tcx>, ) -> Pat<'tcx>

Convert back to a thir::Pat for diagnostic purposes.

pub fn hoist_witness_pat(&self, pat: &WitnessPat<'p, 'tcx>) -> Pat<'tcx>

Convert back to a thir::Pat for diagnostic purposes. This panics for patterns that don’t appear in diagnostics, like float ranges.