rustc_pattern_analysis
Module rustc

Modules

print 🔒

Pattern analysis sometimes wants to print patterns as part of a user-visible diagnostic.

Structs

RevealedTy

A type which has gone through cx.reveal_opaque_ty, i.e. if it was opaque it was replaced by the hidden type if allowed in the current body. This ensures we consistently inspect the hidden types when we should.

RustcPatCtxt

Functions

analyze_match

The entrypoint for this crate. Computes whether a match is exhaustive and which of its arms are useful, and runs some lints.

expand_or_pat 🔒

Recursively expand this pattern into its subpatterns. Only useful for or-patterns.

would_print_as_wildcard 🔒

Returns true if the given pattern would be printed as a wildcard (_).

Type Aliases

Constructor

ConstructorSet

DeconstructedPat

MatchArm

RedundancyExplanation

Usefulness

UsefulnessReport

WitnessPat